comGetObjectInfo {rcom}	R Documentation

Get Information about a COM Objects

Description

This function will return information about the COM object passed as parameter.

Use this function to find out what rcom thinks about the object.

Use this function for debugging purposes to find out why rcom fails.

Not yet implemented!

Usage

comGetObjectInfo(handle)

Arguments

handle 	 COM object (class "COMObject") as returned by e.g. comCreateObject

Value

comGetObject will return a list of all functions and their descriptions

Author(s)

Thomas Baier

[Package rcom version 1.2.4 Index]