comGetVersion {rcom}	R Documentation

Get rcom version information

Description

This function returns the version of the currently installed rcom package as a floating point number.

Use this function to find out which version you are using. The version must be specified when reporting problems.

Usage

comGetVersion()

Arguments

Value

comGetVersion returns the version number in format major.minor.

Author(s)

Thomas Baier

[Package rcom version 1.2.4 Index]