comRegisterRegistry {rcom}	R Documentation

Register the COM server in the registry

Description

This function will register the COM server and its associated type library with the system (registry).

Technically, this will register the type library and add local-server entries for the prog and class ids.

This function has to be called when installing the package. Otherwise, neither the type library will be available (automation clients cannot use the server) nor a mapping from version independent and version-dependant prog-id will be available.

Usage

comRegisterRegistry()

Arguments

Author(s)

Thomas Baier

See Also

comUnregisterRegistry

[Package rcom version 1.2.4 Index]