comRegisterServer {rcom}	R Documentation

Initialize the COM server

Description

This function registers the COM server at runtime. It is called automatically when the package rcom is loaded.

Technically, this will register the class factory in the system, so calls to CoCreateInstance() from client applications will succeed.

Remark: The type library will be loaded on demand later on

comUnregisterServer() is used to unregister the class object (class factory) again.

Usage

comRegisterServer()

Arguments

Author(s)

Thomas Baier

See Also

comUnregisterServer, comRegisterRegistry

[Package rcom version 1.2.4 Index]