comTrace {rcom}	R Documentation

Trace information about SEXP

Description

This function will trace all passed parameters (variables, constants etc) to the R console.

Use this function to find out what rcom thinks about the parameters (e.g. which data types are recognized).

Use this function for debugging purposes to find out why rcom fails.

Usage

comTrace(x)

Arguments

x 	 expression (variable, constant) to trace

Author(s)

Thomas Baier

[Package rcom version 1.2.4 Index]