comUnregisterRegistry {rcom}	R Documentation

Deregister the COM server from the registry

Description

This function will deregister the COM server and its associated type library from the registry/system.

Call this function before uninstalling the package to clean-up the registry.

Usage

comUnregisterRegistry()

Arguments

Author(s)

Thomas Baier

See Also

comRegisterRegistry

[Package rcom version 1.2.4 Index]